# apps/contracts/openapi.json
{
  "info": {
    "title": "Mecatl Studio API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "components": {
    "schemas": {},
    "parameters": {}
  },
  "paths": {
    "/api/v1/auth/session": {
      "get": {
        "operationId": "getAuthSession",
        "responses": {
          "200": {
            "description": "How this deployment identifies callers, and this browser's own state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "mode": {
                      "type": "string",
                      "enum": [
                        "oidc",
                        "static",
                        "none"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "authenticated",
                        "anonymous",
                        "disabled"
                      ]
                    }
                  },
                  "required": [
                    "mode",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "get": {
        "operationId": "startAuthLogin",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "returnTo",
            "in": "query"
          }
        ],
        "responses": {
          "302": {
            "description": "Continue the browser authentication flow.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "required": true
              }
            }
          },
          "400": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/callback": {
      "get": {
        "operationId": "completeAuthLogin",
        "responses": {
          "302": {
            "description": "Continue the browser authentication flow.",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "required": true
              }
            }
          },
          "400": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "logoutAuthSession",
        "responses": {
          "204": {
            "description": "The browser session was cleared."
          },
          "403": {
            "description": "The authentication operation could not be completed.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/health": {
      "get": {
        "operationId": "getHealth",
        "responses": {
          "200": {
            "description": "The Studio BFF is listening.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "service": {
                      "type": "string",
                      "enum": [
                        "mecatl-studio"
                      ]
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  },
                  "required": [
                    "service",
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/runtime": {
      "get": {
        "operationId": "getRuntime",
        "responses": {
          "200": {
            "description": "The connected Mecatl runtime and its capabilities.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "apiMajor": {
                      "type": "integer"
                    },
                    "capabilities": {
                      "type": "object",
                      "properties": {
                        "agents": {
                          "type": "boolean"
                        },
                        "audio": {
                          "type": "boolean"
                        },
                        "bash": {
                          "type": "boolean"
                        },
                        "debugMcp": {
                          "type": "boolean"
                        },
                        "image": {
                          "type": "boolean"
                        },
                        "learnedSkills": {
                          "type": "boolean"
                        },
                        "learningProposals": {
                          "type": "boolean"
                        },
                        "manualCompaction": {
                          "type": "boolean"
                        },
                        "manualDream": {
                          "type": "object",
                          "properties": {
                            "projectMemory": {
                              "type": "object",
                              "properties": {
                                "decide": {
                                  "type": "boolean"
                                },
                                "generate": {
                                  "type": "boolean"
                                },
                                "unavailableReason": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "decide",
                                "generate"
                              ]
                            },
                            "userModel": {
                              "type": "object",
                              "properties": {
                                "decide": {
                                  "type": "boolean"
                                },
                                "generate": {
                                  "type": "boolean"
                                },
                                "unavailableReason": {
                                  "type": "string"
                                }
                              },
                              "required": [
                                "decide",
                                "generate"
                              ]
                            }
                          }
                        },
                        "mcp": {
                          "type": "boolean"
                        },
                        "mcpConnectorStatus": {
                          "type": "boolean"
                        },
                        "memory": {
                          "type": "boolean"
                        },
                        "modelSelection": {
                          "type": "boolean"
                        },
                        "posture": {
                          "type": "string"
                        },
                        "reflection": {
                          "type": "boolean"
                        },
                        "scheduling": {
                          "type": "boolean"
                        },
                        "sessionDebug": {
                          "type": "boolean"
                        },
                        "skills": {
                          "type": "boolean"
                        },
                        "slashCommands": {
                          "type": "boolean"
                        },
                        "soul": {
                          "type": "boolean"
                        },
                        "steer": {
                          "type": "boolean"
                        },
                        "storageCleanup": {
                          "type": "boolean"
                        },
                        "storageHealth": {
                          "type": "boolean"
                        },
                        "storageMigration": {
                          "type": "boolean"
                        },
                        "teams": {
                          "type": "boolean"
                        },
                        "userModel": {
                          "type": "boolean"
                        },
                        "workspaceEnrollment": {
                          "type": "boolean"
                        },
                        "worktrees": {
                          "type": "boolean"
                        }
                      },
                      "required": [
                        "agents",
                        "audio",
                        "bash",
                        "debugMcp",
                        "image",
                        "learnedSkills",
                        "learningProposals",
                        "manualCompaction",
                        "mcp",
                        "mcpConnectorStatus",
                        "memory",
                        "modelSelection",
                        "posture",
                        "reflection",
                        "scheduling",
                        "sessionDebug",
                        "skills",
                        "slashCommands",
                        "soul",
                        "steer",
                        "storageCleanup",
                        "storageHealth",
                        "storageMigration",
                        "teams",
                        "userModel",
                        "workspaceEnrollment",
                        "worktrees"
                      ]
                    },
                    "connection": {
                      "type": "string",
                      "enum": [
                        "connecting",
                        "online",
                        "reconnecting",
                        "offline",
                        "unauthorized",
                        "incompatible"
                      ]
                    },
                    "deployment": {
                      "type": "string"
                    },
                    "features": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    },
                    "mock": {
                      "type": "boolean"
                    },
                    "source": {
                      "type": "string",
                      "enum": [
                        "external",
                        "local"
                      ]
                    }
                  },
                  "required": [
                    "apiMajor",
                    "capabilities",
                    "connection",
                    "features",
                    "mock",
                    "source"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "Sign in before using this deployment.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "The Mecatl runtime is unavailable or not yet negotiated.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "detail": {
                      "type": "string"
                    },
                    "instance": {
                      "type": "string"
                    },
                    "status": {
                      "type": "integer"
                    },
                    "title": {
                      "type": "string"
                    },
                    "type": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "detail",
                    "instance",
                    "status",
                    "title",
                    "type"
                  ]
                }
              }
            }
          }
        }
      }
    }
  }
}